requirements.txt
flask==3.1.0
matplotlib==3.8.0
numpy==1.26.4
pandas==2.2.2
scikit_learn==1.5.1
Jinja2==3.1.4
notebook==7.3.3
pytesseract
Pillow
fpdf